Refactor configuration classes to avoid redundancy
